+++ search-films/src/api/film.constants.ts
@@ -1,0 +1,9 @@
+export const URL_TO_DATA : string = 'https://search.imdbot.workers.dev';
+
+// Ex: https://search.imdbot.workers.dev/?q=Avengers: Endgame
+export const FILM_NAME_PARAMETER = 'q';
+// Ex: https://search.imdbot.workers.dev/?tt=tt4154796
+export const FILM_ID_PARAMETER = 'tt';
+
+// if this flag is true, use mock data
+export const USE_MOCKS = true;