+++ search-films/src/components/Rating/index.tsx
@@ -4,7 +4,7 @@ import { RatingProps } from './Rating.props';
 export default function Rating({value} : RatingProps) {
 	return (
 		<div className={styles['rating']}>
-			<img alt='Иконка рейтинга' src='star-icon.svg'/>
+			<img alt='Иконка рейтинга' src='/star-icon.svg'/>
 			<div className={styles['rating-value']}>{value}</div>
 		</div>
 	);