+++ search-films/src/main.tsx
@@ -1,26 +1,62 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+import { createBrowserRouter, defer, RouterProvider } from 'react-router-dom';
+import { doGetDetails, doSearchFilmDescByName } from './api';
 import { UserProvider } from './context/user.context';
 import './index.css';
 import MenuLayout from './layouts/MenuLayout';
+import { RequireAuth } from './layouts/RequireAuth';
+import SearchLayout from './layouts/SearchLayout';
 import ErrorPage from './pages/ErrorPage/index.js';
 import FavoritesPage from './pages/FavoritesPage/index.js';
 import { LoginPage, MainPage } from './pages/index.js';
+import MoviePage from './pages/MoviePage';
+import SearchPage from './pages/SearchPage';
 
 const router = createBrowserRouter([
 	{
 		path: '/',
-		element: <MenuLayout/>,
+		element: <RequireAuth><MenuLayout/></RequireAuth>,
 		children: [
 			{
 				path: '/',
-				element: <MainPage/>
+				element: <SearchLayout><MainPage/></SearchLayout>
+			},
+			{
+				path: '/movies/:movieName',
+				element: <SearchLayout><SearchPage/></SearchLayout>,
+				errorElement: <ErrorPage/>,
+				loader: async ({ params }) => {
+					if (params.movieName == undefined) {
+						throw new Error('Invalid movie name');
+					}
+					return defer({data: doSearchFilmDescByName(params.movieName)});
+
+				}
+			},
+			{
+				path: '/movie/:movieId',
+				element: <MoviePage/>,
+				errorElement: <ErrorPage/>,
+				loader: async ({ params }) => {
+					if (params.movieId == undefined) {
+						throw new Error('Invalid movie id');
+					}
+					return defer({data: doGetDetails(params.movieId)});
+
+				}
 			},
 			{
 				path: '/favorites',
 				element: <FavoritesPage/>
-			},
+			}
+		]
+	
+	},
+	{
+		path: '/',
+		element: <MenuLayout/>,
+		children: [
 			{
 				path: '/login',
 				element: <LoginPage/>
@@ -28,10 +64,10 @@ const router = createBrowserRouter([
 			{
 				path: '*',
 				element: <ErrorPage/>
-
 			}
 		]
 	}
+	
 ]);
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(