+++ search-films/src/pages/LoginPage/index.tsx
@@ -1,16 +1,19 @@
 import { useContext, useRef } from 'react';
+import { useNavigate } from 'react-router-dom';
 import { Input, Title } from '../../components';
 import { IUserContext, UserContext } from '../../context/user.context';
 
 function LoginPage() {
 	const { saveCurrentUser } = useContext(UserContext) as IUserContext;
-			
+	const navigate = useNavigate();
 	const inputRef = useRef(null);
 
 	const onSubmit = (text: string) => {
 		// if text is not empty set isAuthenticated to true
 		if(text.trim().length > 0) {
 			saveCurrentUser?.(text);
+			navigate('/');
+
 		} else {
 			if (inputRef.current) {
 				(inputRef.current as HTMLElement).focus();